Implement Channel Service
This is an initial commit for the Channel Service. It sets up the required utilities to implement the API functions described in the design document.

`conf.neutron` module supplies a neutron configuration group, following the pattern of the other groups

Two modules (`keystone`, `auth`) are added in the `neutron` package which facilitate authentication to the neutron server. These modules were lifted from [doni](https://github.com/ChameleonCloud/doni).

Several new exception types have been added:
    - `CatalogNotFound`, `KeystoneUnauthorized`, `KeystoneFailure`, and `ConfigInvalid` were added with the Keystone auth modules
    - `MalformedChannel` is used in channel validation if OpenStack returns a malformed channel that was possibly created manually or via a bug in the neutron plugins

A number of new utility functions are introduced to extract information from OpenStack API responses and convert them back and forth between that format and the Channel API format.

The `api.schema` module provides classes, functions, and constants for validating request schema. For the initial release, this will only end up being used in `CreateChannel` and `UpdateChannel`, but this makes API functions that read the request body have an easier time making assumptions about the data that is in those requests, and will also make tunelo easier to extend in the future.

A lot of the utility functions I believe would be made obsolete from a future refactor which will re-organize the internal representation of neutron data into classes. However, for now, they remain.