Add CreateChannel
Implements the CreateChannel API function. Implements functions for creating ports and subnets, as well as resolving channel agent hosts.

At the moment, the code is _very_ branchy and relies a lot on resolving data from dicts with strings. After realizing that OpenStack is extremely inconsistent with the keys they use to represent the same data in their API, it seems obvious that Neutron data and Channel data should be represented by objects to avoid confusion to readers and future programmers.

The code is also fairly sloppy, as it made sense that refactoring it to be a bit neater would be a part of the refactor mentioned above.