Fix neutron subnet name lookup
The Neutron GET api for subnets expects a UUID and cannot interpret
a name properly, so in order to look up by name you have to effectively
search for it. Fortunately it is possible to request a list of subnets
and pass in a name search query.

This also updates how the neutron client is first imported so it is
lazy. It is quite important to wait until as late as possible to initialize
anything that is (1) stored in global state and (2) is configured via
oslo_config. The reason is that oslo_config parsing of conf files happens
quite late, usually after evaluation time, but early in run time (in
the entry point.) If conf is read at eval time, it can be incorrect as the
conf has not yet been properly initialized.